change to doubles

+++ src/main/java/cs455/aqi/q6/Q6Reducer.java
@@ -6,7 +6,6 @@ import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
-import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import java.util.TreeMap;
 import java.util.Map;
@@ -22,12 +21,13 @@ import org.apache.hadoop.mapreduce.lib.output.FileOutputFormat;
 */
 
 public class Q6Reducer extends Reducer<Text, IntWritable, Text, DoubleWritable> {
+    
     @Override
     public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException{
         // sum of all aqi scores
-        long sum = 0;
+        double sum = 0;
         // keeps track of number of items for key
-        int num = 0;
+        double num = 0;
         double avg = 0;
         for(IntWritable val : values){
             num += 1;
